Adjust users location update

+++ persistence/src/main/java/com/lucianthomaz/alpr/alprintegration/persistence/user/PersistentUser.java
@@ -20,8 +20,8 @@ public class PersistentUser {
     private String email;
     private String username;
     private String password;
-    private Float lastKnownLatitude;
-    private Float lastKnownLongitude;
+    private Double lastKnownLatitude;
+    private Double lastKnownLongitude;
     private String lastModified;
     private String deviceFcmToken;
 }
